+++ src/lib/components/index.ts
@@ -31,7 +31,7 @@ export { default as RectClipPath } from './RectClipPath.svelte';
 export { default as Sankey } from './Sankey.svelte';
 export { default as Text } from './Text.svelte';
 export { default as Threshold } from './Threshold.svelte';
-export { default as Tooltip } from './Tooltip.svelte';
+export { default as TooltipContext } from './TooltipContext.svelte';
 export { default as TooltipContainer } from './TooltipContainer.svelte';
 export { default as TooltipItem } from './TooltipItem.svelte';
 export { default as TooltipSeparator } from './TooltipSeparator.svelte';
